requirements.txt
django==1.4.2
django-dajaxice==0.5.5
raven==4.1.1
django-endless-pagination==1.1
numpy==1.8.1
pyPdf==1.13
django-fixture-magic==0.0.3
requests==2.2.1
